resolved incorrect file use, and created round class
