+++ lib/round.rb
@@ -1,21 +1,3 @@
-require 'minitest/autorun'
-require 'minitest/pride'
-require 'pry'
-require './lib/card'
-require './lib/turn'
-require './lib/deck'
-require './lib/round'
-
-class RoundTest < MiniTest::Test
-  def test_it_exists
-    card_1 = Card.new("What is the capital of Alaska?", "Juneau", :Geography)
-    card_2 = Card.new("The Viking spacecraft sent back to Earth photographs and reports about the surface of which planet?", "Mars", :STEM)
-    card_3 = Card.new("Describe in words the exact direction that is 697.5° clockwise from due north?", "North north west", :STEM)
-    cards = [card_1, card_2, card_3]
-    deck = Deck.new(cards)
-    round = Round.new(deck)
-
-    assert_instance_of Round, round
-  end
+class Round
 
 end
